requirements.dev.txt
coloredlogs
flake8
flask-socketio
pylint
pytest
pytest-cov
python-dotenv
